requirements.txt
accelerate==0.29.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.1
anyio==4.3.0
attrs==23.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
boilerpy3==1.0.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
commlib-py==0.10.6
confection==0.1.4
contourpy==1.2.1
coverage==5.5
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
datasets==2.18.0
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
eyed3==0.9.7
filelock==3.13.4
filetype==1.2.0
Flask==3.0.3
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
google-ai-generativelanguage==0.6.2
google-ai-haystack==1.0.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
haystack-ai==2.1.1
haystack-bm25==1.0.2
hiredis==2.1.1
hmmlearn==0.3.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
HyperPyYAML==1.2.2
idna==3.6
imbalanced-learn==0.12.2
imblearn==0.0
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==2.4
k2==0.3.0
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-text-splitters==0.0.1
langcodes==3.4.0
langsmith==0.1.45
language_data==1.2.0
lazy-imports==0.3.1
llama-index==0.10.37
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.37.post1
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-groq==0.1.4
llama-index-llms-openai==0.1.19
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.22
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.3
llamaindex-py-client==0.1.19
llvmlite==0.42.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newsapi-python==0.2.7
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.28.1
openai-whisper==20231117
orjson==3.10.0
outcome==1.3.0.post0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.2.2
pause==0.3
pika==1.3.1
pillow==10.3.0
plotly==5.22.0
pluggy==1.5.0
ply==3.11
posthog==3.5.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyAudioAnalysis==0.3.14
pycparser==2.22
pydantic==1.10.4
pydantic_core==2.16.3
pydub==0.25.1
pymongo==4.6.3
pyparsing==3.1.2
pypdf==4.2.0
PySocks==1.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.1.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
scrypt==0.8.24
selenium==4.20.0
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.29
srsly==2.4.8
striprtf==0.0.26
sympy==1.12
telegram==0.0.1
tenacity==8.2.3
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.7.0
uritemplate==4.1.1
urllib3==2.2.1
URLObject==2.4.3
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
